Create Controllers and Fake data
- Create ProductController
- Create TagController
- Fix Product model by add Eloquent\Relations to use it
- Fix field type to unsignedBigInteger in media table migration
- Add seeders class to DatabaseSeeder
- Add Products and Products Tags Faker data in EntrustSeeder
- Add Logic to Generate Fake data for ProductSeeder
- Add Logic to Generate Fake data for ProductsImagesSeeder
- Add Logic to Generate Fake data for ProductsTagSeeder
- Add Logic to Generate Fake data for TagSeeder
- Add products and tags routes in web.php
- Create Products and Tags Views for backend

+++ database/migrations/2021_08_31_143640_create_media_table.php
@@ -19,8 +19,8 @@ class CreateMediaTable extends Migration
             $table->string('file_type');
             $table->string('file_size');
             $table->string('file_status');
-            $table->unsignedTinyInteger('file_sort')->default(0);
-            $table->unsignedTinyInteger('mediable_id');
+            $table->unsignedBigInteger('file_sort')->default(0);
+            $table->unsignedBigInteger('mediable_id');
             $table->string('mediable_type');
             $table->timestamps();
         });
